@@ -1,68 +1,97 @@
 {
-  "name": "kville-scheduler",
-  "private": true,
-  "dependencies": {
-    "@rails/webpacker": "3.5",
-    "axios": "^0.18.0",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-lodash": "^3.2.11",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-1": "^6.24.1",
-    "cross-env": "^5.0.5",
-    "history": "^4.7.2",
-    "husky": "^3.1.0",
-    "moment": "^2.22.2",
-    "prop-types": "^15.6.1",
-    "query-string": "^6.2.0",
-    "randomstring": "^1.1.5",
-    "react": "^16.3.2",
-    "react-big-calendar": "^0.20.2",
-    "react-count-down": "^1.1.0",
-    "react-dom": "^16.3.2",
-    "react-forecast": "git+https://github.com/mafuvadze/react-forecast.git",
-    "react-redux": "^5.0.6",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.3.1",
-    "react-router-redux": "^5.0.0-alpha.9",
-    "react-timeago": "^4.1.9",
-    "react-twitter-widgets": "^1.7.1",
-    "react-youtube": "^7.6.0",
-    "redux": "^3.7.2",
-    "redux-auth-wrapper": "^2.0.2",
-    "redux-devtools": "^3.4.1",
-    "redux-devtools-dock-monitor": "^1.1.3",
-    "redux-devtools-log-monitor": "^1.4.0",
-    "redux-persist": "^5.6.6",
-    "redux-search": "^2.4.0",
-    "redux-thunk": "^2.2.0",
-    "semantic-ui-calendar-react": "^0.12.1",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.81.2"
-  },
-  "devDependencies": {
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-stage-0": "^6.24.1",
-    "webpack-dev-server": "2.11.2"
-  },
-  "scripts": {
-    "start": "foreman start -f Procfile.dev",
-    "setup": "./bin/setup.sh",
-    "prod": "foreman start -f Procfile.prod",
-    "clean:branch": "./bin/clean-branch.sh",
-    "webpacker": "ruby bin/webpack-dev-server",
-    "flush": " docker-compose run app ./bin/db-flush.sh",
-    "dev:setup": "./bin/dev.sh",
-    "dev:start": "docker-compose run -p '5000:5000' -p '3035:3035' app yarn start",
-    "dev:bash": "docker-compose run app bash",
-    "dev:console": "docker-compose run app rails c",
-    "dev:db": "docker-compose run app rails db",
-    "dev:db:migrate": "docker-compose run app rails db:migrate",
-    "dev:db:seed": "docker-compose run app rails db:seed",
-    "prod:start": "./bin/prod-start.sh",
-    "prod:build": "./bin/build-production.sh",
-    "staging:build": "./bin/build-staging.sh",
-    "dev:clean-start": "./bin/clean-start-server.sh"
-  }
+    "name": "kville-scheduler",
+    "private": true,
+    "dependencies": {
+        "@babel/core": "^7.19.3",
+        "@babel/plugin-proposal-object-rest-spread": "^7.19.3",
+        "@babel/preset-env": "^7.19.3",
+        "@babel/preset-react": "^7.18.6",
+        "@babel/preset-typescript": "^7.18.6",
+        "@types/react-router-dom": "^5.3.3",
+        "autoprefixer": "^10.4.12",
+        "axios": "^1.1.2",
+        "babel-loader": "^8.2.5",
+        "connected-react-router": "^6.9.3",
+        "cross-env": "^7.0.3",
+        "css-loader": "^6.7.1",
+        "file-loader": "^6.2.0",
+        "history": "4.10.1",
+        "husky": "^8.0.1",
+        "mini-css-extract-plugin": "^2.6.1",
+        "moment": "^2.22.2",
+        "postcss-import": "^15.0.0",
+        "postcss-loader": "^7.0.1",
+        "prop-types": "^15.6.1",
+        "query-string": "^7.1.1",
+        "randomstring": "^1.1.5",
+        "react": "^18.2.0",
+        "react-big-calendar": "^1.5.0",
+        "react-count-down": "^1.1.0",
+        "react-dom": "^18.2.0",
+        "react-forecast": "git+https://github.com/mafuvadze/react-forecast.git",
+        "react-redux": "^8.0.4",
+        "react-router": "^6.4.2",
+        "react-router-dom": "^6.4.2",
+        "react-timeago": "^7.1.0",
+        "react-twitter-widgets": "^1.7.1",
+        "react-youtube": "^10.0.0",
+        "redux": "^4.2.0",
+        "redux-auth-wrapper": "^3.0.0",
+        "redux-devtools": "^3.4.1",
+        "redux-devtools-dock-monitor": "^1.1.3",
+        "redux-devtools-log-monitor": "^2.1.0",
+        "redux-persist": "^6.0.0",
+        "redux-search": "^2.4.0",
+        "redux-thunk": "^2.2.0",
+        "sass": "^1.55.0",
+        "sass-loader": "^13.1.0",
+        "semantic-ui-calendar-react": "^0.15.3",
+        "semantic-ui-css": "^2.4.1",
+        "semantic-ui-react": "^2.1.3",
+        "style-loader": "^3.3.1",
+        "webpack": "^5.74.0",
+        "webpack-cli": "^4.10.0",
+        "webpack-remove-empty-scripts": "^1.0.1"
+    },
+    "devDependencies": {
+        "babel-polyfill": "^6.26.0"
+    },
+    "scripts": {
+        "start": "foreman start -f Procfile.dev",
+        "setup": "./bin/setup.sh",
+        "prod": "foreman start -f Procfile.prod",
+        "clean:branch": "./bin/clean-branch.sh",
+        "flush": " docker-compose run app ./bin/db-flush.sh",
+        "dev:setup": "./bin/dev.sh",
+        "dev:start": "docker-compose run -p 5000:5000 -p 3035:3035 app yarn start",
+        "dev:bash": "docker-compose run app bash",
+        "dev:console": "docker-compose run app rails c",
+        "dev:db": "docker-compose run app rails db",
+        "dev:db:migrate": "docker-compose run app rails db:migrate",
+        "dev:db:seed": "docker-compose run app rails db:seed",
+        "prod:start": "./bin/prod-start.sh",
+        "prod:build": "./bin/build-production.sh",
+        "staging:build": "./bin/build-staging.sh",
+        "dev:clean-start": "./bin/clean-start-server.sh",
+        "build": "webpack --config ./webpack.config.js"
+    },
+    "babel": {
+        "presets": [
+            [
+                "@babel/preset-env",
+                {
+                    "targets": {
+                        "browsers": "> 1%",
+                        "uglify": true
+                    }
+                }
+            ],
+            "@babel/preset-react",
+            "@babel/preset-typescript"
+        ],
+        "plugins": [
+            "@babel/syntax-dynamic-import",
+            "@babel/plugin-proposal-object-rest-spread"
+        ]
+    }
 }
