@@ -1,12 +1,12 @@
 // Hack around
 // https://github.com/ReactTraining/react-router/issues/5072
 import { connect } from 'react-redux';
-import Switch from 'react-router-dom/Switch';
+import { Routes } from 'react-router-dom';
 
 const mapStateToProps = state => ({
   location: state.router.location,
 });
 
-const ConnectedSwitch = connect(mapStateToProps)(Switch);
+const ConnectedSwitch = connect(mapStateToProps)(Routes);
 
 export default ConnectedSwitch;
